tilføjet Service til Api og kan ses på hjemmesidden + ændring af navn af apihelper klassen
